+++ src/components/combo_box/types.ts
@@ -27,7 +27,7 @@ export type EuiComboBoxOptionOption<
   Omit<ButtonHTMLAttributes<HTMLButtonElement>, 'value'> & {
     isGroupLabelOption?: boolean;
     label: string;
-    id?: string;
+    key?: string;
     options?: Array<EuiComboBoxOptionOption<T>>;
     value?: T;
   };